Reconfigured structure and added body parser for POST
